Added CheckerTest12()
Added the function CheckerTest12(), which tests that function isPrime and isPrimeRepunit() give the expected result.
We test the 18th first Super-poulet number with more than 2 different prime factors. Because these numbers are not prime, we expect the functions isPrime and isPrimeRempunit to return false for each of them.
I also added CheckerTest10() and CheckerTest11() to the main function as they were mising. I suspect their contributors forgot to add them.